requirements.txt
seaborn==0.11.1
calplot==0.1.7.3
pydeck==0.6.1
altair==4.1.0
streamlit==0.79.0
geopandas==0.6.1
matplotlib==3.3.4
joblib==1.0.1
spacy_streamlit==1.0.0rc1
pandas==1.2.2
nltk==3.5
spacy_nightly==3.0.0rc5
numpy==1.19.2
scipy==1.6.1
wordcloud==1.8.1
statsmodels==0.12.2
Pillow==8.2.0
scikit_learn==0.24.2
spacy==3.0.6
